Use size_t for row and column variables.
	Rather than int, use size_t for row and column variables,
	since they are essentially sizes.  size_t is unsigned, so we
	get the full range of values representable by the underlying
	type, rather than only half (we can't have a matrix with
	negative dimensions).  Since size_t is commonly an unsigned
	long int, this also means we can represent vastly larger
	matrices.

	The struct matrix is changed to use size_t internally.  All
	functions accepting row or column parameters have had the
	types of their parameters updated.  All loop variables which
	iterate over rows or columns are also updated.  The printf
	call in mtx_print has had its format specifiers for the row
	and column count updated as well.

	An additional benefit, to be taken advantage of in a later
	commit, is that mtx_new uses the row and column parameters as
	its arguments to the malloc functions, which expect size_t
	parameters.  While there aren't compiler complaints about
	passing an int where a size_t is expected (that I encountered,
	at least), this is more correct, anyway.  Some internal
	variables used for rows or columns also had their types
	updated; e.g., the mtx_get_{row,col}_vector functions call
	mtx_new, so they use the appropriate type as well.

	* lib/matrix.h: Use size_t for row and column variables.
	* lib/mtx_add.c: Likewise.
	* lib/mtx_copy.c: Likewise.
	* lib/mtx_delete.c: Likewise.
	* lib/mtx_get_col_vector.c: Likewise.
	* lib/mtx_get_row_vector.c: Likewise.
	* lib/mtx_identity.c: Likewise.
	* lib/mtx_multiply.c: Likewise.
	* lib/mtx_new.c: Likewise.
	* lib/mtx_print.c: Likewise.
	* lib/mtx_rref.c: Likewise.
	* lib/mtx_transpose.c: Likewise.
	* lib/mtx_vector_norm2.c: Likewise.

+++ lib/matrix.h
@@ -10,8 +10,8 @@
 /* Struct.  */
 
 struct matrix {
-	int rows, cols;
-	double** data;
+	size_t rows, cols;
+	double **data;
 };
 
 
@@ -19,7 +19,7 @@ struct matrix {
 
 // New Matrix
 // Make a new matrix, elements initialized to zero
-struct matrix *mtx_new (int rows, int columns);
+struct matrix *mtx_new (size_t rows, size_t columns);
 
 
 // Delete all the data of a matrix
@@ -28,7 +28,7 @@ void mtx_delete (struct matrix *m);
 
 
 // Make an identity matrix of dimension dim
-struct matrix *mtx_identity (int dim);
+struct matrix *mtx_identity (size_t dim);
 
 
 //Copy the matrix
@@ -64,13 +64,13 @@ void mtx_print (struct matrix *m);
 // Get a column vector.
 // col == true -> a column vector using column [col] and rows [low] to [high]
 // includes low row, not  high
-struct matrix *mtx_get_col_vector (struct matrix *m, int low, int high, int col);
+struct matrix *mtx_get_col_vector (struct matrix *m, size_t low, size_t high, size_t col);
 
 
 // Get a row vector.
 // a row vector using row [row] and columns [low] to [high]
 // includes low column, not high
-struct matrix *mtx_get_row_vector (struct matrix *m, int low, int high, int row);
+struct matrix *mtx_get_row_vector (struct matrix *m, size_t low, size_t high, size_t row);
 
 
 // Norm of a vector, squared
