+++ lib/mtx_new.c
@@ -4,14 +4,14 @@
 #include "matrix.h"
 
 struct matrix *
-mtx_new (int rows, int columns)
+mtx_new (size_t rows, size_t columns)
 {
 	struct matrix *m = (struct matrix *) calloc (1, sizeof (struct matrix));
 
 	m->rows = rows;
 	m->cols = columns;
 	m->data = (double **) calloc (rows, sizeof (double *));
-	int i;
+	size_t i;
 	for (i = 0;  i < rows; i++ )
 		m->data[i] = (double *) calloc (columns, sizeof (double));
